refactor : 모델명 변경
