+++ local/src/main/java/com/kky/pokedex/local/AppDatabase.kt
@@ -3,9 +3,9 @@ package com.kky.pokedex.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import com.kky.pokedex.local.dao.LikeDao
-import com.kky.pokedex.local.model.Like
+import com.kky.pokedex.local.model.LikeEntity
 
-@Database(entities = [Like::class], version = 1)
+@Database(entities = [LikeEntity::class], version = 1)
 abstract class AppDatabase: RoomDatabase() {
     abstract fun dao(): LikeDao
 }
