+++ local/src/main/java/com/kky/pokedex/local/dao/LikeDao.kt
@@ -5,16 +5,16 @@ import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.kky.pokedex.local.model.Like
+import com.kky.pokedex.local.model.LikeEntity
 
 @Dao
 interface LikeDao {
     @Query("SELECT * FROM `like`")
-    suspend fun getLikes(): List<Like>
+    suspend fun getLikes(): List<LikeEntity>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insertLike(like: Like)
+    suspend fun insertLike(like: LikeEntity)
 
     @Delete
-    suspend fun deleteLike(like: Like)
+    suspend fun deleteLike(like: LikeEntity)
 }
