+++ network/src/main/java/com/kky/pokedex/network/model/PokemonFormResponse.kt
@@ -2,18 +2,7 @@ package com.kky.pokedex.network.model
 
 import com.google.gson.annotations.SerializedName
 
-data class Pokemon(
-    val id: Int,
-    val names: List<PokemonName>,
-    @SerializedName("flavor_text_entries") val descriptions: List<FlavorText>
-)
-
-data class PokemonName(
-    val name: String,
-    val language: Info
-)
-
-data class PokemonForm(
+data class PokemonFormResponse(
     @SerializedName("sprites") val images: PokemonImage,
     val types: List<PokemonTypes>
 )
