+++ network/src/main/java/com/kky/pokedex/network/model/PokemonResponse.kt
@@ -1,0 +1,14 @@
+package com.kky.pokedex.network.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Pokemon(
+    val id: Int,
+    val names: List<PokemonName>,
+    @SerializedName("flavor_text_entries") val descriptions: List<FlavorText>
+)
+
+data class PokemonName(
+    val name: String,
+    val language: Info
+)
